Refactor restaurant business logic into RestaurantService
